Implemented SQL parameters assigned to values, This is needed as some preleminary work to fix WHERE clauses and implement WHERE IN
