+++ src/Traits/Where.php
@@ -153,6 +153,8 @@ trait Where
         return $this;
     }
 
+    abstract protected function addStatementParameter(?string $parameter, $value): void;
+
     abstract protected function isValidSqlName(string $name): bool;
 
     /**
